feat: connect to database via typeorm

+++ src/cart/services/cart.service.ts
@@ -1,59 +1,76 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Cart as CartRepo } from '../../database/entities/cart.entity';
+import { Cart } from '../../database/entities/cart.entity';
 
 import { v4 } from 'uuid';
 
-import { Cart } from '../models';
+// import { Cart } from '../models';
 
 @Injectable()
 export class CartService {
   constructor(
-    @InjectRepository(CartRepo) private readonly cartRepository: Repository<Cart>,
+    @InjectRepository(Cart)
+    private readonly cartRepository: Repository<Cart>,
   ) {}
-  // private userCarts: Record<string, Cart> = {};
+  private userCarts: Record<string, Cart> = {};
 
-  findByUserId(user_id: string) {
-      return this.cartRepository.findOneBy({ user_id });
+  findByUserId(id: string) {
+    return this.cartRepository.find({
+      where: { user_id: id },
+      relations: ['items'],
+      // loadRelationIds: true,
+    });
   }
 
   createByUserId(userId: string) {
-    // const id = v4();
+    const id = v4();
     const userCart = {
+      id,
       user_id: userId,
-    } as unknown as Cart;
+      items: [],
+    };
 
-    this.cartRepository.save(userCart)
+    console.log('userCart', userCart);
+
+    this.cartRepository.save(userCart);
 
     return userCart;
   }
 
   async findOrCreateByUserId(userId: string): Promise<Cart> {
     const userCart = await this.findByUserId(userId);
+    console.log(userCart);
 
-    if (userCart) {
-      return userCart;
+    if (userCart[0]) {
+      return userCart[0];
     }
 
-    return this.createByUserId(userId);
+    const newUser = (await this.createByUserId(userId)) as unknown as Cart;
+    return newUser;
   }
 
-  // updateByUserId(userId: string, { items }): Cart {
-  //   const { id, ...rest } = this.findOrCreateByUserId(userId);
+  async updateByUserId(userId: string, { items }): Promise<Cart> {
+    const { id, ...rest } = await this.findOrCreateByUserId(userId);
 
-  //   const updatedCart = {
-  //     id,
-  //     ...rest,
-  //     items: [...items],
-  //   };
+    const updatedCart = {
+      id,
+      ...rest,
+      items: [...items],
+    };
 
-  //   this.userCarts[userId] = { ...updatedCart };
+    const updatedCartEntity = await this.cartRepository.preload(updatedCart);
 
-  //   return { ...updatedCart };
-  // }
+    await this.cartRepository.save(updatedCartEntity);
+
+    this.userCarts[userId] = { ...updatedCart };
+
+    return { ...updatedCart };
+  }
 
   removeByUserId(userId): void {
-    this.cartRepository.delete(userId);
+    console.log('userId', userId);
+
+    this.cartRepository.delete({ user_id: userId });
   }
 }
